refactor: env validation typescript and common package
