+++ packages/api/src/router/invite/checks.ts
@@ -1,6 +1,6 @@
 import { TRPCError } from "@trpc/server";
 
-import { env } from "@homarr/auth/env.mjs";
+import { env } from "@homarr/auth/env";
 
 export const throwIfCredentialsDisabled = () => {
   if (!env.AUTH_PROVIDERS.includes("credentials")) {
