+++ packages/api/src/router/test/group.spec.ts
@@ -1,7 +1,7 @@
 import { describe, expect, test, vi } from "vitest";
 
 import type { Session } from "@homarr/auth";
-import * as env from "@homarr/auth/env.mjs";
+import * as env from "@homarr/auth/env";
 import { createId, eq } from "@homarr/db";
 import { groupMembers, groupPermissions, groups, users } from "@homarr/db/schema";
 import { createDb } from "@homarr/db/test";
