+++ packages/auth/package.json
@@ -11,7 +11,7 @@
     "./client": "./client.ts",
     "./server": "./server.ts",
     "./shared": "./shared.ts",
-    "./env.mjs": "./env.mjs"
+    "./env": "./env.ts"
   },
   "main": "./index.ts",
   "types": "./index.ts",
