+++ packages/auth/providers/check-provider.ts
@@ -1,6 +1,6 @@
 import type { SupportedAuthProvider } from "@homarr/definitions";
 
-import { env } from "../env.mjs";
+import { env } from "../env";
 
 export const isProviderEnabled = (provider: SupportedAuthProvider) => {
   // The question mark is placed there because isProviderEnabled is called during static build of about page
