+++ packages/common/env.ts
@@ -2,6 +2,8 @@ import { randomBytes } from "crypto";
 import { createEnv } from "@t3-oss/env-nextjs";
 import { z } from "zod";
 
+import { shouldSkipEnvValidation } from "./src/env-validation";
+
 const errorSuffix = `, please generate a 64 character secret in hex format or use the following: "${randomBytes(32).toString("hex")}"`;
 
 export const env = createEnv({
@@ -23,6 +25,5 @@ export const env = createEnv({
   runtimeEnv: {
     SECRET_ENCRYPTION_KEY: process.env.SECRET_ENCRYPTION_KEY,
   },
-  skipValidation:
-    Boolean(process.env.CI) || Boolean(process.env.SKIP_ENV_VALIDATION) || process.env.npm_lifecycle_event === "lint",
+  skipValidation: shouldSkipEnvValidation(),
 });
