+++ packages/common/package.json
@@ -9,7 +9,8 @@
     "./types": "./src/types.ts",
     "./server": "./src/server.ts",
     "./client": "./src/client.ts",
-    "./env.mjs": "./env.mjs"
+    "./env": "./env.ts",
+    "./env-validation": "./src/env-validation.ts"
   },
   "typesVersions": {
     "*": {
@@ -30,7 +31,8 @@
     "dayjs": "^1.11.13",
     "next": "15.1.4",
     "react": "19.0.0",
-    "react-dom": "19.0.0"
+    "react-dom": "19.0.0",
+    "zod": "^3.24.1"
   },
   "devDependencies": {
     "@homarr/eslint-config": "workspace:^0.2.0",
