+++ packages/db/configs/mysql.config.ts
@@ -1,6 +1,6 @@
 import type { Config } from "drizzle-kit";
 
-import { env } from "../env.mjs";
+import { env } from "../env";
 
 export default {
   dialect: "mysql",
